refactor(push): extract push-open handling from Klaviyo into KlaviyoPushOpenHandler
